refactor(match): recover toon edge vector ownership

+++ tools/match/scratches/dot_vector/scratch.cpp
@@ -1,6 +1,6 @@
 #include "vector3.h"
 
-double Vector3::dot_vector(const Vector3* rhs)
+float Vector3::dot_vector(const Vector3* rhs)
 {
     return rhs->z * z + rhs->y * y + rhs->x * x;
 }
